Added unit tests for BasicFormulas

+++ NanoXLSX/Address.cs
@@ -121,8 +121,8 @@ namespace NanoXLSX
         /// <returns>-1 if the other address is greater, 0 if equal and 1 if smaller</returns>
         public int CompareTo(Address other)
         {
-            long thisCoordinate = Column * Worksheet.MAX_ROW_NUMBER + Row;
-            long otherCoordinate = other.Column * Worksheet.MAX_ROW_NUMBER + other.Row;
+            long thisCoordinate = (long)Column * (long)Worksheet.MAX_ROW_NUMBER + Row;
+            long otherCoordinate = (long)other.Column * (long)Worksheet.MAX_ROW_NUMBER + other.Row;
             return thisCoordinate.CompareTo(otherCoordinate);
         }
